requirements.txt
boto3>=1.34.0
pandas>=2.2.0
streamlit>=1.29.0
plotly>=5.18.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0